requirements.txt
fastapi==0.104.1
psycopg2==2.9.6
pydantic~=1.10.13
bcrypt==4.1.1
Jinja2==3.1.2
python_jose==3.3.0
python-multipart==0.0.6
Requests==2.31.0
starlette==0.27.0
uvicorn==0.25.0
python-jose~=3.3.0
psycopg2-binary~=2.9.9
python-dotenv~=1.0.0
numpy~=1.26.4
pandas~=1.5.3
pgvector~=0.4.1
pydub~=0.25.1
librosa~=0.10.2.post1
scipy~=1.10.1
crepe~=0.0.16